Added problem

+++ CodeForces_Problems/EasyMathProblems/RobotCleaner.cpp
@@ -1,24 +1,24 @@
 #include<bits/stdc++.h>
 using namespace std;
 
-#define vi vector<int>
-#define vii vector<int,int>
-#define rep(a,b) for(int i = a;i<b;i++)
-#define pi pair<int,int>
+#define vi vector<ll>
+#define vii vector<ll,ll>
+#define rep(a,b) for(ll i = a;i<b;i++)
+#define pi pair<ll,ll>
 
 typedef long long ll;
 
 int main(){
    ios_base::sync_with_stdio(false);cin.tie(NULL);cout.tie(NULL);
 
-   int t;
+   ll t;
    cin>>t;
 
    while(t--){
-      int n, m, xr, yr, xd, yd;
+      ll n, m, xr, yr, xd, yd;
       cin>>n>>m>>xr>>yr>>xd>>yd;
 
-      int dr = 1, dh = 1, ans = 0;
+      ll dr = 1, dh = 1, ans = 0;
 
       while(true){
          if(xr == xd || yr == yd) break;
